requirements.txt
Django==5.0.6
numpy==1.26.4
djangorestframework
djangorestframework-simplejwt
psycopg[binary]
celery[redis]
diffusers
torch
Pillow
python-decouple
transformers
accelerate
huggingface-hub>=0.16.0
python-dotenv
redis
django-admin-interface
django-colorfield
